Fixes to make it work with dx run
