+++ src/cgppindel.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
-# cgppindel 1.0.0
+# cgppindel 1.1.0
 
 set -e -x -o pipefail
 
+# Install all python packages from this app
+sudo -H python3 -m pip install --no-index --no-deps packages/*
 
 main() {
 
@@ -64,18 +66,18 @@ main() {
     # Add Allele frequency (AF) and Read depth (DP) onto cgppindel output file
     vcf=$(find out/cgppindel_output -type f -name "*.vcf.gz")
     basename=$(basename "$(basename "$vcf" .gz)" .vcf)
-    python tsv_file_generator.py -v "$vcf" 
+    python3 /tsv_file_generator.py -v "$vcf" 
     bgzip annots.tsv
     tabix -s1 -b2 -e2 annots.tsv.gz
-    bcftools annotate -a annots.tsv.gz -h annots.hdr -c CHROM,POS,ID,REF,ALT,+FORMAT/AF,+FORMAT/DP \
-    "$vcf" > "$basename.af.vcf.gz"
+    bcftools annotate -a annots.tsv.gz -h /annots.hdr -c CHROM,POS,ID,REF,ALT,+FORMAT/AF,+FORMAT/DP \
+    "$vcf" > "$basename.af.vcf"
 
     # Move vcf and index in specified runfolder to enable downstream use
     mv out/cgppindel_output/*.flagged.vcf.gz out/output_vcf
     mv out/cgppindel_output/*.flagged.vcf.gz.tbi out/vcf_index
 
     # Move vcf file with VAF in speified folder
-    mv $basename.af.vcf.gz out/output_vcf_with_vaf
+    mv $basename.af.vcf out/output_vcf_with_vaf
 
     # Move all .out and .err to a temporary folder
     mv out/cgppindel_output/logs/*.err temp_logs
